WIP: minimal image & interrpose without manifest
This creates a minimal visible image surface suitable for compositing
with filesystem as described by manifest. There is a shadow content
directory containing basic Photon 5.0 binaries that can be invoked
via interpose mechanism. LD_LIBRARY_PATH is modified appropriately.

There is an interpose binary present that uses a hardcoded mapping of
/x -> /.shadow/content/x instead of any sophistication from manifest.
This is enough to support shell and CLI util invocation.

Total image size works out at 14MiB for the minimal image, 42MiB for
shadow content, and 3.7 for meta (manifest and interpose).