Support interpose on execution in container backed mocks
This adds the basic pattern for intercepting executions inside a container backing.

This consists of:
* binary that replaces the target in the container image and will call out to vcsim for instruction prior to invoking the original binary
* configuration file indicating where to talk to vcism
* vcsim parts to accept an incoming connection from the interpose binary and deserialize the attempted invocation.
* methods on host and containers in vcism to manipulate the config.

The communication protocol between mock and vcsim is SSH. This was chosen as it's primary design intent was to support invocation of commands on remote systems, and relay of input/output/status.

This does NOT currently provide any of the mechanisms by which a test case can register interest in interpositions and intercept calls.