Use eth0 as a fallback nic name for serving interpose
The current code looks for the docker bridge to get an IP for the
server that will be accessible from containers.
However with podman, if there's no powered on container then the
bridge doesn't exist. This falls back to using eth0 IP if there's
no extant bridge NIC.

We don't use 0.0.0.0 because we need to provide an actual IP to
the client for connecting, and cannot use loopback as we've got
different net namespaces.